Refactor SkipGramTest, add logger and improve tests description
Refactoring of SkipGramTest.java was done to improve its readability and maintainability. Logger was added instead of using System.out.println for debugging purposes which makes it a cleaner and more reliable approach. Changed the names of test methods to be more descriptive and accurate. Added proper assertions in tests to check the exact result rather than generalized logging.