Moved AliasSampling to common package and added dot method in VectorUtils
The AliasSampling class and its test have been moved to the 'common' package, due to its wide usage across the application. This reflects the intent to make it a common utility class which can be used by many components of the application. Furthermore, the 'dot' method previously residing in Word2Vec.java has been added to VectorUtils.java to make this method available to all who need to calculate the dot product of two vectors. By making these changes, we hope to improve code clarity and increase reusability of functions within our codebase.